requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
blessed==1.20.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.2
django-background-tasks==1.2.8
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-dirtyfields==1.9.7
django-filter==25.1
django-mailer==2.3.2
django-oauth-toolkit==3.0.1
django-picklefield==3.3
django-q2==1.8.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-keycloak==1.0.1
ecdsa==0.19.1
idna==3.10
jinxed==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.5.3
lockfile==0.12.2
logger==1.4
Markdown==3.7
oauthlib==3.2.2
paramiko==3.5.1
pathlib==1.0.1
ping3==4.0.8
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
python-keycloak-client==0.2.3
pytz==2025.2
redis==3.5.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
schedule==1.2.2
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
